add support of PDEF in ctl
