+++ README.md
@@ -4,7 +4,7 @@
 
 
 ## 1. Introduction
-The Grid Analysis and Display System ([GrADS](http://cola.gmu.edu/grads/) or [OpenGrADS](http://www.opengrads.org/)) is a widely used software for easy access, manipulation, and visualization of earth science data.  It uses a descriptor (or control) file with a suffix `.ctl` to  describe a raw binary 4D dataset.  The `ctl` file is similar to the header information of a [NetCDF](https://www.unidata.ucar.edu/software/netcdf/docs/file_structure_and_performance.html) file, containing all the information about dimensions, attributes, and variables except for the variable data.
+The Grid Analysis and Display System ([GrADS](http://cola.gmu.edu/grads/) or [OpenGrADS](http://www.opengrads.org/)) is a widely used software for easy access, manipulation, and visualization of earth science data.  It uses a [descriptor (or control) file with a suffix `.ctl`](http://cola.gmu.edu/grads/gadoc/descriptorfile.html) to  describe a raw binary 4D dataset.  The `ctl` file is similar to the header information of a [NetCDF](https://www.unidata.ucar.edu/software/netcdf/docs/file_structure_and_performance.html) file, containing all the information about dimensions, attributes, and variables except for the variable data.
 
 This python package [`xgrads`](https://github.com/miniufo/xgrads) is designed for parse and read the `.ctl` file commonly used by [GrADS](http://cola.gmu.edu/grads/).  Right now it can parse various kinds of `.ctl` files.  However, only the commonly used raw binary 4D datasets can be read using [`dask`](https://dask.org/) and return as a [`xarray.Dataset`](http://xarray.pydata.org/en/stable/)  Other types of binary data, like `dtype` is `station` or`grib`, may be supported in the future.
 
@@ -13,7 +13,7 @@ This python package [`xgrads`](https://github.com/miniufo/xgrads) is designed fo
 ### 2.1 Parse a `.ctl` file
 Parsing a `.ctl` file is pretty simple using the following code:
 ```python
-from xgrads.core import CtlDescriptor
+from xgrads import CtlDescriptor
 
 ctl = CtlDescriptor(file='test.ctl')
 
@@ -46,7 +46,7 @@ print(ctl)
 ### 2.2 Read binary data into a `xarray.Dataset`
 Reading a `.ctl` related binary data file is also pretty simple using the following code:
 ```python
-from xgrads.core import open_CtlDataset
+from xgrads import open_CtlDataset
 
 dset = open_CtlDataset('test.ctl')
 
@@ -61,7 +61,7 @@ Then you have the `dset` as a `xarray.Dataset`.  This is similar to [`xarray.ope
 ### 2.3 Convert a GrADS dataset to a NetCDF dataset
 With the above functionality, it is easy to convert a `.ctl` ([GrADS](http://cola.gmu.edu/grads/)) dataset to a `.nc` ([NetCDF](https://www.unidata.ucar.edu/software/netcdf/docs/file_structure_and_performance.html)) dataset:
 ```python
-from xgrads.core import open_CtlDataset
+from xgrads import open_CtlDataset
 
 open_CtlDataset('input.ctl').to_netcdf('output.nc')
 ```
