+++ xgrads/__init__.py
@@ -1,3 +1,4 @@
 # -*- coding: utf-8 -*-
-from .core import CtlDescriptor, open_CtlDataset
+from .core import CtlDescriptor
+from .io import open_CtlDataset
 __version__ = "0.1.0"
